requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.9
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cleanup==9.0.0
django-environ==0.13.0
django-extensions==4.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
django-tinymce==5.0.0
gunicorn==25.1.0
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
pillow==12.0.0
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.4
